Initial Python Wrapping Setup (#34)
* Fixed cmake variable defined check.

* Also removed buildDependencies target from `all`
* Updated README
* Fixed github action for windows and macos

* Added step for starting workflow in ViennaPS repository

* Added workflow for checking coding style

* Changed unsigned to signed type in omp for loop to calm msvc

* CMake minimum version bump

* Restructured CMake configuration and added initial skeleton for python
bindings

* Applied formatting script

* Formatted cmake scripts

* Updated workflows

* Updated workflows

* Updated workflows

* Added option for static embree build. Introduced setup.py

* Dependencies dir can now be set by the user

* Updated Static build option

* Extended gitignore, removed static option from setup.py

* Update LICENSE date

* Added VIENNARAY_DEPENDENCIES_DIR override

+++ include/rayTrace.hpp
@@ -189,7 +189,7 @@ public:
     case rayNormalizationType::MAX: {
       auto maxv = *std::max_element(flux.begin(), flux.end());
 #pragma omp parallel for
-      for (size_t idx = 0; idx < flux.size(); ++idx) {
+      for (int idx = 0; idx < flux.size(); ++idx) {
         flux[idx] *= (totalDiskArea / diskArea[idx]) / maxv;
       }
       break;
@@ -202,7 +202,7 @@ public:
                               : mNumberOfRaysFixed;
       NumericType normFactor = sourceArea / numTotalRays;
 #pragma omp parallel for
-      for (size_t idx = 0; idx < flux.size(); ++idx) {
+      for (int idx = 0; idx < flux.size(); ++idx) {
         flux[idx] *= normFactor / diskArea[idx];
       }
       break;
@@ -220,7 +220,7 @@ public:
            "Unequal number of points in smoothFlux");
     auto oldFlux = flux;
 #pragma omp parallel for
-    for (size_t idx = 0; idx < mGeometry.getNumPoints(); idx++) {
+    for (int idx = 0; idx < mGeometry.getNumPoints(); idx++) {
       auto neighborhood = mGeometry.getNeighborIndicies(idx);
       for (auto const &nbi : neighborhood) {
         flux[idx] += oldFlux[nbi];
